requirements.txt
asgiref==3.3.0
certifi==2020.11.8
chardet==3.0.4
Django==3.1.3
idna==2.10
journal==0.4.0
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
urllib3==1.26.2
